Move blockwatch into the ethereum package
