+++ core/core.go
@@ -7,9 +7,9 @@ import (
 	"context"
 	"time"
 
-	"github.com/0xProject/0x-mesh/blockwatch"
 	"github.com/0xProject/0x-mesh/db"
 	"github.com/0xProject/0x-mesh/ethereum"
+	"github.com/0xProject/0x-mesh/ethereum/blockwatch"
 	"github.com/0xProject/0x-mesh/meshdb"
 	"github.com/0xProject/0x-mesh/p2p"
 	"github.com/0xProject/0x-mesh/zeroex"
