+++ ethereum/blockwatch/fake_client.go
