+++ zeroex/orderwatch/watcher.go
@@ -8,9 +8,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/0xProject/0x-mesh/blockwatch"
 	"github.com/0xProject/0x-mesh/constants"
 	"github.com/0xProject/0x-mesh/db"
+	"github.com/0xProject/0x-mesh/ethereum/blockwatch"
 	"github.com/0xProject/0x-mesh/meshdb"
 	"github.com/0xProject/0x-mesh/zeroex"
 	"github.com/ethereum/go-ethereum/common"
